+++ horqrux/utils.py
@@ -40,7 +40,7 @@ def _dagger(operator: Array) -> Array:
 
 def _unitary(generator: Array, theta: float) -> Array:
     return (
-        jnp.cos(theta / 2) * jnp.eye(2, dtype=jnp.complex128) - 1j * jnp.sin(theta / 2) * generator
+        jnp.cos(theta / 2) * jnp.eye(2, dtype=jnp.complex64) - 1j * jnp.sin(theta / 2) * generator
     )
 
 
@@ -48,14 +48,14 @@ def _jacobian(generator: Array, theta: float) -> Array:
     return (
         -1
         / 2
-        * (jnp.sin(theta / 2) * jnp.eye(2, dtype=jnp.complex128) + 1j * jnp.cos(theta / 2))
+        * (jnp.sin(theta / 2) * jnp.eye(2, dtype=jnp.complex64) + 1j * jnp.cos(theta / 2))
         * generator
     )
 
 
 def _controlled(operator: Array, n_control: int) -> Array:
     n_qubits = int(log2(operator.shape[0]))
-    control = jnp.eye(2 ** (n_control + n_qubits), dtype=jnp.complex128)
+    control = jnp.eye(2 ** (n_control + n_qubits), dtype=jnp.complex64)
     control = control.at[-(2**n_qubits) :, -(2**n_qubits) :].set(operator)
     return control
 
@@ -70,7 +70,7 @@ def product_state(bitstring: str) -> Array:
         A state corresponding to 'bitstring'.
     """
     n_qubits = len(bitstring)
-    space = jnp.zeros(tuple(2 for _ in range(n_qubits)), dtype=jnp.complex128)
+    space = jnp.zeros(tuple(2 for _ in range(n_qubits)), dtype=jnp.complex64)
     space = space.at[tuple(map(int, bitstring))].set(1.0)
     return space
 
@@ -129,8 +129,8 @@ def overlap(state: Array, projection: Array) -> Array:
 def uniform_state(
     n_qubits: int,
 ) -> Array:
-    state = jnp.ones(2**n_qubits, dtype=jnp.complex128)
-    state = state / jnp.sqrt(jnp.array(2**n_qubits, dtype=jnp.complex128))
+    state = jnp.ones(2**n_qubits, dtype=jnp.complex64)
+    state = state / jnp.sqrt(jnp.array(2**n_qubits, dtype=jnp.complex64))
     return state.reshape([2] * n_qubits)
